Former-commit-id: 3626c57c40758079c905c6b2baf3ca083bc320b5 [formerly b6af57b411b9ec4c9fa3ba83714e67adf8c15db1] [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614] [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee]] [formerly 3626c57c40758079c905c6b2baf3ca083bc320b5 [formerly b6af57b411b9ec4c9fa3ba83714e67adf8c15db1] [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614] [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee]] [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly 29677326c7a0b9dcf115b06d8cb2e0645a2c7614 [formerly be88c9494ecedbdb140a4995c6feec6b29c4cfee] [formerly 6a1714a7b00a5510b395a91ab3d4e4170651907a [formerly f65e8a7253c65b5d59e78ed57e3d73bf01b2e06f]]]]]
Former-commit-id: 6a1714a7b00a5510b395a91ab3d4e4170651907a
Former-commit-id: 09819032a19ca24f7ce9a24d39b3e3d51adef500 [formerly a11d23c5050a27e81cfdaca00c335e8e5c1cd72c] [formerly e280c293be772e3b76431146949849c8b59106f3] [formerly 76ce22b10999b668d133b7331eae90c7f6254db1 [formerly 119ca9e825f2d313a0116934bf8546a845a3fdd3] [formerly e280c293be772e3b76431146949849c8b59106f3 [formerly 44ecd4408b0a495b887d8514b0d31c9c7be47f74]]]
Former-commit-id: 3c6d4cb7bc43689cbe5c7f5a44205dd14a005bf8 [formerly 2dfa9af138bcd404a0418985f7b7715288c66892] [formerly 9a44f4f5fde9f96ae2d0d1aac7aa3844cc5b3c00 [formerly 6af9e88e9374794fdb017013e2bcc49bbaf728f2]]
Former-commit-id: 9a44f4f5fde9f96ae2d0d1aac7aa3844cc5b3c00
Former-commit-id: 82bb918f2fe097a81cb9f1f10fac429ec344d8d5

+++ edexOsgi/com.raytheon.edex.plugin.sfcobs/src/com/raytheon/edex/plugin/sfcobs/decoder/synoptic/MobileSynopticDecoder.java
@@ -39,6 +39,7 @@ import com.raytheon.uf.edex.decodertools.core.IDecoderConstants;
  * Date         Ticket#    Engineer    Description
  * ------------ ---------- ----------- --------------------------
  * 20070928            391 jkorman     Initial Coding.
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * </pre>
  * 
  * @author jkorman
@@ -49,9 +50,9 @@ public class MobileSynopticDecoder extends AbstractSynopticDecoder {
     // The logger
     private Log logger = LogFactory.getLog(getClass());
 
-    private Double mobileLatitude = null;
+    private Float mobileLatitude = null;
 
-    private Double mobileLongitude = null;
+    private Float mobileLongitude = null;
 
     private Integer mobileQuadrant = null;
 
@@ -173,7 +174,7 @@ public class MobileSynopticDecoder extends AbstractSynopticDecoder {
         if (matchElement(element, "99\\d{3}")) {
             Integer lat = getInt(element, 2, 5);
             if (lat != null) {
-                mobileLatitude = lat.doubleValue() / 10.0;
+                mobileLatitude = lat.floatValue() / 10.0f;
             }
         }
     }
@@ -192,7 +193,7 @@ public class MobileSynopticDecoder extends AbstractSynopticDecoder {
         if (matchElement(element, "[1357]((0\\d{3})|(1(([0-7]\\d{2})|(800))))")) {
             Integer lon = getInt(element, 2, 5);
             if (lon != null) {
-                mobileLongitude = lon.doubleValue() / 10.0;
+                mobileLongitude = lon.floatValue() / 10.0f;
             }
             mobileQuadrant = getInt(element, 0, 1);
         }
@@ -207,8 +208,8 @@ public class MobileSynopticDecoder extends AbstractSynopticDecoder {
         if ((mobileLatitude != null) && (mobileLongitude != null)
                 && (mobileQuadrant != null)) {
             if ((mobileLatitude >= 0) && (mobileLongitude >= 0)) {
-                double lat = 0;
-                double lon = 0;
+                float lat = 0;
+                float lon = 0;
                 switch (mobileQuadrant) {
                 case 1: {
                     lat = 1;
